+++ PaoBu/Libraries/Categories/NSString+Helpers.m
@@ -111,9 +111,8 @@
     return result;
 }
 //将Int转成Str
-NSString* stringWithInt(int number){
-    
-    return [NSString stringWithFormat:@"%d",number];
+NSString* stringWithInt(NSInteger number){
+    return [NSString stringWithFormat:@"%ld",(long)number];
 }
 //将double转成Str
 NSString* stringWithDouble(double number){
@@ -178,7 +177,7 @@ NSString* stringWithFloat(float number){
 
     const char* str = [self UTF8String];
     unsigned char result[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(str, strlen(str), result);
+    CC_MD5(str, (unsigned int)strlen(str), result);
     NSMutableString *ret = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH*2];//
 
     for(int i = 0; i<CC_MD5_DIGEST_LENGTH; i++) {
